initial release: JWTDecode
