+++ app-android/src/main/java/dev/sdkforge/jwt/decode/android/MainActivity.kt
@@ -1,0 +1,24 @@
+package dev.sdkforge.jwt.decode.android
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.ui.Modifier
+import dev.sdkforge.jwt.decode.app.App
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(
+        savedInstanceState: Bundle?,
+    ) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            App(
+                modifier = Modifier
+                    .fillMaxSize(),
+            )
+        }
+    }
+}
