--- app-android/src/main/java/dev/sdkforge/template/android/MainActivity.kt
@@ -1,24 +1,0 @@
-package dev.sdkforge.template.android
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.ui.Modifier
-import dev.sdkforge.template.app.App
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(
-        savedInstanceState: Bundle?,
-    ) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            App(
-                modifier = Modifier
-                    .fillMaxSize(),
-            )
-        }
-    }
-}
