+++ app-shared/build.gradle.kts
@@ -21,5 +21,5 @@ kotlin {
 }
 
 android {
-    namespace = "dev.sdkforge.template.app"
+    namespace = "dev.sdkforge.jwt.decode.app"
 }
