--- app-shared/src/commonMain/kotlin/dev/sdkforge/template/app/App.kt
@@ -1,37 +1,0 @@
-package dev.sdkforge.template.app
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import dev.sdkforge.template.core.currentPlatform
-
-@Composable
-fun App(
-    modifier: Modifier = Modifier,
-) = ApplicationTheme {
-    Surface(
-        modifier = modifier,
-        color = MaterialTheme.colorScheme.background,
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(
-                space = 8.dp,
-                alignment = Alignment.CenterVertically,
-            ),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Text(
-                text = "Platform name: ${currentPlatform.name}",
-            )
-            Text(
-                text = "Platform version: ${currentPlatform.version}",
-            )
-        }
-    }
-}
