+++ app-shared/src/iosMain/kotlin/dev/sdkforge/jwt/decode/app/ComposeAppViewController.kt
@@ -1,0 +1,20 @@
+package dev.sdkforge.jwt.decode.app
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.window.ComposeUIViewController
+import kotlin.experimental.ExperimentalObjCName
+
+@OptIn(ExperimentalObjCName::class)
+@Suppress("FunctionName")
+@ObjCName("create")
+fun ComposeAppViewController() = ComposeUIViewController(
+    configure = {
+        enforceStrictPlistSanityCheck = false
+    },
+) {
+    App(
+        modifier = Modifier
+            .fillMaxSize(),
+    )
+}
