--- app-shared/src/iosMain/kotlin/dev/sdkforge/template/app/ComposeAppViewController.kt
@@ -1,20 +1,0 @@
-package dev.sdkforge.template.app
-
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.window.ComposeUIViewController
-import kotlin.experimental.ExperimentalObjCName
-
-@OptIn(ExperimentalObjCName::class)
-@Suppress("FunctionName")
-@ObjCName("create")
-fun ComposeAppViewController() = ComposeUIViewController(
-    configure = {
-        enforceStrictPlistSanityCheck = false
-    },
-) {
-    App(
-        modifier = Modifier
-            .fillMaxSize(),
-    )
-}
