+++ gradle.properties
@@ -12,8 +12,8 @@ android.nonTransitiveRClass=true
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
 org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
 #Publishing
-publishing.group=dev.sdkforge.template
+publishing.group=dev.sdkforge.jwt.decode
 publishing.version=0.0.1
 publishing.owner=SDKForge
-publishing.repository=template-sdk
+publishing.repository=JWTDecode
 publishing.description=A modern Kotlin Multiplatform SDK template for building cross-platform libraries and applications.
