+++ settings.gradle.kts
@@ -43,7 +43,7 @@ dependencyResolutionManagement {
     }
 }
 
-rootProject.name = "SDKForgeTemplate"
+rootProject.name = "SDKForge-JWTDecode"
 
 include(":app-android")
 include(":app-shared")
