+++ shared/build.gradle.kts
@@ -47,5 +47,5 @@ kotlin {
 }
 
 android {
-    namespace = "dev.sdkforge.template"
+    namespace = "dev.sdkforge.jwt.decode"
 }
